backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
pymysql==1.1.1
cryptography==42.0.8
pydantic==2.8.2
pydantic-settings==2.3.4
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
alembic==1.13.2
python-multipart==0.0.9
